Introduce a setting for a more privacy aware lastlogin
If we log the exact date of the last login, this can potentially be
used to link it against other patterns of recorded data. If we log
only the last month of the login, we should have obfuscated any
usage patterns, but still be able to decide whether an account is
still used or not.